+++ subpackages/cli/source/odood/cli/commands/assembly.d
@@ -12,7 +12,7 @@ private import colored;
 private import commandr: Argument, Option, Flag, ProgramArgs;
 private import thepath: Path;
 
-private import odood.lib.assembly: Assembly;
+private import odood.lib.assembly: Assembly, ASSEMBLY_VERSION_PATH;
 private import odood.lib.assembly.exception: OdoodAssemblyNothingToCommitException;
 private import odood.lib.project: Project;
 private import odood.utils.addons.addon: OdooAddon;
@@ -37,6 +37,7 @@ class CommandAssemblyInit: OdoodCommand {
             project.initializeAssembly();
         else
             project.initializeAssembly(parseGitURL(args.option("repo")));
+            project.assembly.get.link();
     }
 }
 
@@ -95,6 +96,8 @@ class CommandAssemblySync: OdoodCommand {
             null, "push", "Automatically push changes if needed."));
         this.add(new Option(
             null, "push-to", "Name of branch to push changes to."));
+        this.add(new Flag(
+            null, "changelog", "Generate changelog for assembly."));
     }
 
     public override void execute(ProgramArgs args) {
@@ -106,13 +109,24 @@ class CommandAssemblySync: OdoodCommand {
         // Do the sync
         project.assembly.get.sync();
 
+        if (args.flag("changelog"))
+            project.assembly.get.generateChangelog();
+
         // Commit changes if requested (usually useful in CI flows)
         if (args.flag("commit") || args.flag("push") || args.option("push-to")) {
             enforce!OdoodCLIException(
                 project.assembly.get.repo.getChangedFiles(path_filters: [":(exclude)dist"], staged: false).length == 0,
                 "Assembly Sync: There are unexpected changes in assembly. Please, handle it manually.");
             enforce!OdoodCLIException(
-                project.assembly.get.repo.getChangedFiles(path_filters: [":(exclude)dist"], staged: true).length == 0,
+                project.assembly.get.repo.getChangedFiles(
+                    path_filters: [
+                        ":(exclude)dist",
+                        ":(exclude)%s".format(ASSEMBLY_VERSION_PATH),
+                        ":(exclude)CHANGELOG.md",
+                        ":(exclude)CHANGELOG.latest.md",
+                    ],
+                    staged: true
+                ).length == 0,
                 "Assembly Sync: There are unexpected staged changes in assembly. Please, handle it manually.");
 
             if (project.assembly.get.repo.getChangedFiles(path_filters: ["dist"], staged: true)) {